% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfunctions.R, R/giscourse.R
\name{connect}
\alias{connect}
\title{Make a connection to the gis_course data base quickly.}
\usage{
connect(db = "gis_course")

connect(db = "gis_course")
}
\arguments{
\item{No}{parameters needed}

\item{No}{parameters needed}
}
\value{
returns the connection

returns the connection
}
\description{
Make a connection to the gis_course data base quickly.

Make a connection to the gis_course data base quickly.
}
\examples{
conn<-connect()

conn<-connect()

}
