% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{dem}
\alias{dem}
\title{Get's DEM around point}
\usage{
dem(x = "Arne,Dorset", y = 0, dist = 0, z = 12)
}
\arguments{
\item{z}{}
}
\description{
z is set to 12 by default (12 m)
Use value of 14 for highest resolution (3m)
}
\examples{



}
