% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{dquery}
\alias{dquery}
\title{Distance query
Finds all geometries for a query lying within a distance of a
place, found through Geocoding}
\usage{
dquery(place = "Hengistbury head", dist = 1000,
  query = "select * from ph_v2_1")
}
\arguments{
\item{query}{}
}
\description{
Distance query
Finds all geometries for a query lying within a distance of a
place, found through Geocoding
}
\examples{
dquery() \%>\% qtm()
}
