% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{emap}
\alias{emap}
\title{Quickly draw some}
\usage{
emap(place = "Bournemouth", write = TRUE, table = "my_edits")
}
\arguments{
\item{place}{Defaults to Bournemouth. Geocoded.}

\item{write}{Defaults to TRUE.Writes to the course data base}

\item{table}{Name of the table in the data base to write to}
}
\value{
A geometry
}
\description{
Quickly draw some
}
