% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{getwclim}
\alias{getwclim}
\title{Getting World Clim data}
\usage{
getwclim(place = "Bournemouth", res = 2.5)
}
\arguments{
\item{place}{Defaults to Bournemouth}

\item{res}{Resolution defaults to 10}
}
\value{
Returns a data frame
}
\description{
Pulls out worldclim prec, tmin and tmax from the layers
}
