% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{hansen_wms}
\alias{hansen_wms}
\title{Add Hansen's deforestation maps to tropical areas in WMS format
This is for visualisation only}
\usage{
hansen_wms(m)
}
\arguments{
\item{m}{}
}
\description{
Add Hansen's deforestation maps to tropical areas in WMS format
This is for visualisation only
}
\examples{
qmap("Sumatra") \%>\% hansen_wms()
}
