% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{insol}
\alias{insol}
\title{Direct beam insolation for day of the year}
\usage{
insol(dtm, day = "13/10/19")
}
\arguments{
\item{day}{}
}
\description{
Calculated using SAGA
Returns a raster pointing to a temporary file
Save locally if you need it to be permanent
}
\examples{

dtm<-mkdem()
sol<-insol(dtm)

}
