% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{landcover}
\alias{landcover}
\title{Gets the landcover for a site}
\usage{
landcover(x = "Arne,Dorset", y = 0, dist = 0)
}
\arguments{
\item{dist}{}
}
\value{
Landcover clipped to the site
}
\description{
Uses the same rules for input as mksite.
Provide either a string or lat long coordinates.
}
\examples{

landcover("Arne,Dorset"dist=1000)

}
