% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{make_db}
\alias{make_db}
\title{Super user make a new data base
Use with great care. Will wipe out prexisting data base!}
\usage{
make_db(db = "gis_course2")
}
\arguments{
\item{db}{}
}
\description{
Super user make a new data base
Use with great care. Will wipe out prexisting data base!
}
