% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{mkpnt}
\alias{mkpnt}
\title{Make a point}
\usage{
mkpnt(x = "Wareham, Dorset", y = 0)
}
\arguments{
\item{x}{If x is a string then the Geocoding is used to find the centre}

\item{y}{If x and y are provided as numbers they are assumed to be lon and lat}
}
\value{
Sf point in British National grid
}
\description{
This is a quick utility function that feeds into a workflow
It takes either a string or two numbers
If a string is provided the position is based on geocoding
If x and y are provided they are assumed to be lon lat, (or lat lon)
As this is just for the UK we can switch if they are typed in the wrong order
by testing if x > y-
}
\examples{
mkpnt(x="Wareham, Dorset")
mkpnt(x=-1,y=52)


}
