% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{mksite}
\alias{mksite}
\title{Make a site frame}
\usage{
mksite(x = "Arne,Dorset", y = 0, dist = 0, bx = FALSE)
}
\arguments{
\item{dist}{}
}
\description{
Same rules for x and y as in mkpnt
If a distance is provided then it is used as a buffer around the point
If a distance is not provided then the os5km grid square is used.
}
