% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{os5km}
\alias{os5km}
\title{OS 5km grid square}
\usage{
os5km(x = "Wareham,Dorset", y = 0)
}
\arguments{
\item{y}{}
}
\description{
Uses same input rules as mkpnt as it feeds into it.
Provide either some text or lat long coordinates
}
\examples{
os5km("Wareham,Dorset")


}
