% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{osm}
\alias{osm}
\title{Find osm roads and paths around a site}
\usage{
osm(x = "Arne,Dorset", y = 0, dist = 0)
}
\arguments{
\item{dist}{}
}
\description{
Find osm roads and paths around a site
}
\examples{

}
