% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{phab_choose}
\alias{phab_choose}
\title{Choose a type of priority habitat}
\usage{
phab_choose(place = "Bournemouth", phab = "heath", dist = 5000)
}
\arguments{
\item{dist}{}
}
\description{
Give a place, a radius and some words
}
