% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{qmap}
\alias{qmap}
\title{Quckly form a map using geocoding to find a place}
\usage{
qmap(place = "Bournemouth")
}
\arguments{
\item{place}{}
}
\value{
A mapview map
}
\description{
Quckly form a map using geocoding to find a place
}
