% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{sconnect}
\alias{sconnect}
\title{Make super user connection}
\usage{
sconnect(db = "gis_course")
}
\arguments{
\item{db}{Default db is gis_course}
}
\description{
Use with great care, as this gives permsissions to drop tables.
}
\examples{
conn<-sconnect()
}
