% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{wldiag}
\alias{wldiag}
\title{Walter and Leith diagram}
\usage{
wldiag(df)
}
\arguments{
\item{df}{}
}
\description{
Produces a simple WL diagram
Takes the df from getwlim
}
\examples{
 getwclim() \%>\% wldiag()


}
