% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querys.R
\name{xyquery}
\alias{xyquery}
\title{Query within a distance from X and Y}
\usage{
xyquery(x = -1.95, y = 50.667, dist = 1000,
  query = "select * from ph_v2_1")
}
\arguments{
\item{query}{}
}
\description{
X is Lon, Y is Lat
Default query get's priority habitat
}
