% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{connect}
\alias{connect}
\title{Make a connection to the gis_course data base quickly.}
\usage{
connect()
}
\arguments{
\item{No}{parameters needed}
}
\value{
returns the connection
}
\description{
Make a connection to the gis_course data base quickly.
}
\examples{
conn<-connect()

}
