% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{connect}
\alias{connect}
\title{Quick connect}
\usage{
connect(db = "gis_course")
}
\arguments{
\item{No}{parameters needed}
}
\value{
Just returns the connection
}
\description{
Make a connection to the gis_course data base quickly.
}
\examples{
conn<-connect()

}
