% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giscourse.R
\name{sconnect}
\alias{sconnect}
\title{Make super user connection
Use with care}
\usage{
sconnect(db = "gis_course", pwd = "docker")
}
\arguments{
\item{db}{Default is gis_course}
}
\description{
Make super user connection
Use with care
}
